Add the direction of the vote to the toast
As jonas-w mentioned in #161 the voting direction (upvote, downvote or
vote removing) isn't mentioned in the toast. To give better feedback of
the action taken this commit now includes them. Since the information
was already available in the method it was quite straight forward to
implement it.

(cherry picked from commit 053779c68c89bc3f568bfb367d228e36853bd5b0)